Crate data_chain

Data Chain

This crate holds three modules that together allow nodes in a decentralised network to manage data and history of data and nodes in a container that in itself is decentralised.

The modules here include the ability to cryptographically lock history of data and nodes in a chain. There also exists the mechnism o persist these chains over sessions along with a copy of the data itself (the ChuinkStore). The data representation is via the Data module.

Example

Basic usage

TBD

Reexports

pub use chain::{DataChain, NodeBlock, create_link_descriptor};

pub use data::{ImmutableData, MAX_BYTES, PlainData, StructuredData};

Modules

chain
A block is a type that contains a BlockIdentifier and a Proof. These can be data blocks or links

chunk_store
Persistant store on disk of Keys and large values TODO should not be public

data
Data types

error
Error types for this crate

Enums

BlockIdentifier
Data identifiers for use in a data Chain. The hash of each data type is available to ensure there is no confusion over the validity of any data presented by this chain

Data
Data types handled in a SAFE

DataIdentifier
An identifier to address a data chunk.