Module data_chain::block_identifier

Identify the variant parts of a block, for links this is the Digest of the hash of that group.

Enums

BlockIdentifier
Dummy data identifiers for this crate

Type Definitions

Ledger
Ledger type (delee or keep)

LinkDescriptor
Represents the xored close group for the new grou on churn etc. This is signed by each group member.

SdName
structured Data name