Struct data_chain::chain::data_chain::DataChain

pub struct DataChain { /* fields omitted */ }

Created by holder of chain, can be passed to others as proof of data held. This object is verifiable if : The last validation contains the majority of current close group OR on network restart the nodes all must try and restart on previous names. They can continue any validation of the holder of a chain. This requires nodes to always restart as last ID and if there was no restart they are rejected at vault level. If there was a restart then the nodes should validate and continue. N:B this means all nodes can use a named directory for data store and clear if they restart as a new id. This allows clean-up of old data cache directories.

Methods

impl DataChain

fn create_in_path(path: PathBuf, group_size: usize) -> Result<DataChain>

Create a new chain backed up on disk Provide the directory to create the files in

fn from_path(path: PathBuf, group_size: usize) -> Result<DataChain, Error>

Open from existing directory

fn from_blocks(blocks: Vec<Block>, group_size: usize) -> DataChain

Create chain in memory from vector of blocks

fn write(&self) -> Result<(), Error>

Write current data chain to supplied path

fn write_to_new_path(&mut self, path: PathBuf) -> Result<(), Error>

Write current data chain to supplied path

fn unlock(&self)

Unlock the lock file

fn validate_ownership(&mut self, my_group: &[PublicKey]) -> bool

Nodes always validate a chain before accepting it Validation takes place from start of chain to now. Also confirm we can accept this chain, by comparing our current group with the majority of the last known link This method will NOT purge

fn add_vote(&mut self, block: Vote) -> Option<BlockIdentifier>

Add a vote received from a peer Uses lazy accumulation If block becomes valid, then it is returned

fn chain(&self) -> &Vec<Block>

getter

fn find(&self, block_identifier: &BlockIdentifier) -> Option<&Block>

find a block (user required to test for validity)

fn find_name(&self, name: &[u8; 32]) -> Option<&Block>

find block by name from top (only first occurrence)

fn remove(&mut self, data_id: &BlockIdentifier)

Remove a block, will ignore Links

fn retain<F>(&mut self, pred: F) where F: FnMut(&Block) -> bool

Retains only the blocks specified by the predicate.

fn clear(&mut self)

Clear chain

fn contains(&self, block_identifier: &BlockIdentifier) -> bool

Check if chain contains a particular identifier

fn position(&self, block_identifier: &BlockIdentifier) -> Option<usize>

Return position of block identifier

fn insert(&mut self, index: usize, block: Block)

Inserts an element at position index within the chain, shifting all elements after it to the right. Will not validate this block!

Panics

Panics if index is greater than the chains length.

fn validate_block(&mut self, block: &mut Block) -> bool

Validates an individual block. Will get latest link and confirm all signatures were from last known valid group.

fn prune(&mut self)

Removes all invalid blocks, does not confirm chain is valid to this group.

fn len(&self) -> usize

Total length of chain

fn valid_len(&self) -> usize

Number of valid blocks

fn blocks_len(&self) -> usize

number of valid data blocks

fn links_len(&self) -> usize

number of valid links

fn is_empty(&self) -> bool

Contains no blocks that are not valid

fn all_links(&self) -> Vec<Block>

Returns all links in chain Does not perform validation on links

fn valid_data(&mut self) -> Vec<Block>

Validates and returns all links in chain

fn valid_links(&mut self) -> Vec<Block>

Validates and returns all links in chain

fn valid_links_at_block_id(&mut self, block_id: &BlockIdentifier) -> Vec<Block>

Validates and returns all valid links in chain 4 before and after target

fn mark_blocks_valid(&mut self)

Mark all links that are valid as such.

fn merge_chain(&mut self, chain: &mut DataChain)

Merge any blocks from a given chain