Struct data_chain::chain::node_block::NodeBlock

pub struct NodeBlock { /* fields omitted */ }

If data block then this is sent by any group member when data is Put, Post or Delete. If this is a link then it is sent with a churn event. A Link is a nodeblock that each member must send each other in times of churn. These will not accumulate but be ManagedNode to ManagedNode messages in the routing layer

Methods

impl NodeBlock

fn new(pub_key: &PublicKey,
secret_key: &SecretKey,
data_identifier: BlockIdentifier)
-> Result<NodeBlock, Error>

Create a Block (used by nodes in network to send to holders of DataChains)

fn identifier(&self) -> &BlockIdentifier

Getter

fn proof(&self) -> &Proof

Getter

fn validate(&self) -> bool

validate signed correctly

fn validate_detached(&self, identifier: &BlockIdentifier) -> bool

validate signed correctly