Module data_chain::data

Data types Immutable data (name == hash of content) and Structured data (constant name with editable contents and ownership changes) this type is signed by at least one owner and can be a ledger type (never deleted)

Data types

These data types fall into three categories

Immutable : This data type has fixed content and self validates. The name is derived from the hash of the content.

Structured : This is owned data and reflects a location in the address space that contains content that may be altered by the owner(s).

Plain : This is data with no fixed content or name. It is currently unused in SAFE

Reexports

pub use data::immutable_data::ImmutableData;

pub use data::plain_data::PlainData;

pub use data::structured_data::{MAX_BYTES, StructuredData};

Modules

immutable_data
Data that will not change it's contents

plain_data
Flexible and insecure data type with no fixed invarients. The name and content are not related.

structured_data
Data that will retain it's name but allow dynamic content or transfer of ownership

Enums

Data
Data types handled in a SAFE

DataIdentifier
An identifier to address a data chunk.