Struct data_chain::data_chain::DataChain

pub struct DataChain {
    // some fields omitted
}

Created by holder of chain, can be passed to others as proof of data held. This object is verifiable if : The last validation contains the majority of current close group OR on network restart the nodes all must try and restart on previous names. They can continue any validation of the holder of a chain. This requires nodes to always restart as last ID and if there was no restart they are rejected at vault level. If there was a restart then the nodes should validate and continue. N:B this means all nodes can use a named directory for data store and clear if they restart as a new id. This allows clean-up of old data cache directories.

Methods

impl DataChain

fn new() -> DataChain

Create a new chain with no elements yet.

fn validate(&mut self, my_group: &[PublicKey]) -> bool

Nodes always validate a chain before accepting it Validation takes place from start of chain to now. Also confirm we can accept this chain, by comparing our current group with the majority of the last known link This method will purge all note yet valid blocks

fn add_node_block(&mut self, block: NodeBlock) -> Result<(), Error>

Add a nodeblock received from a peer We do not validate the block, it may be out of order This is a case of lazy accumulation

fn prune(&mut self)

Remove invalid blocks and links from chain

fn len(&self) -> usize

Total length of chain

fn blocks_len(&self) -> usize

number of blocks

fn links_len(&self) -> usize

number of links

fn is_empty(&self) -> bool

Contains no blocks that are not valid

fn valid(&mut self, data_id: BlockIdentifier)

valid a block (will not valid a link)

fn get_last_link(&self) -> Option<&Block>

Should equal the current common_close_group

fn get_all_links(&self) -> DataChain

Validate and return all links in chain Does not perform validation on links

fn validate_block(&self, block: &Block) -> Result<(), Error>

Validate an individual block. Will get latest link and confirm all signatures were from last known group. Majority of signatures is confirmed.