Struct data_chain::mmap::FileDataChain

pub struct FileDataChain {
    // some fields omitted
}

A memory map backed data chain

Methods

impl FileDataChain

fn create(&self, path: String, len: u64) -> Result<FileDataChain>

Open or create a FIleDataChain

fn open(path: &File) -> Result<FileDataChain>

Open an existing DataChain for read/write mode

fn open_path<P: AsRef<Path>>(path: P) -> Result<FileDataChain>

Open a new memory map from the path given.

fn len(&self) -> usize

Returns the size in byte of the memory map.

If it is a range, the size is the size of the range.

fn range(&self, offset: usize, len: usize) -> FileDataChain

Slice this memory map to a new offset and len.

Panics

If the new range is outside the bounds of the map

unsafe fn write(&mut self, data: &[u8]) -> Result<usize>

Write date to map

Safety

This method currently unsafe (rust 0.9)