Module data_chain::node_block

A node block is a partial block, sent by group members to each other to create a Block

Structs

NodeBlock
If data block then this is sent by any group member when data is Put, Post or Delete. If this is a link then it is sent with a churn event. A Link is a nodeblock that each member must send each other in times of churn. These will not accumulate but be ManagedNode to ManagedNode messages in the routing layer

NodeBlockProof
Proof as provided by a close group member

Functions

create_link_descriptor
Descriptor is the xored group members starting with base of 0000..:32 This process is faster than hash and means group can be unordered which is beneficial under heavy churn and out of order links being sent.