Struct data_chain::secured_data::SecuredData

pub struct SecuredData { /* fields omitted */ }

API for data based operations.

Methods

impl SecuredData

fn create_in_path(path: PathBuf,
max_disk_space: u64,
group_size: usize)
-> Result<SecuredData, Error>

Construct new data container

fn from_path(path: PathBuf,
max_disk_space: u64,
group_size: usize)
-> Result<SecuredData, Error>

Open an existing container from path

fn clear_disk(&self, path: &Path) -> Result<(), Error>

remove all disk based data

fn chain(&self) -> Arc<Mutex<DataChain>>

Access to DataChain

fn add_vote(&mut self, nb: Vote) -> Option<BlockIdentifier>

Add a Vote from another node If block is valid will return BlockIdentifier

fn has_data(&self, data_id: &DataIdentifier) -> bool

Do we have the data on disk.

fn get(&self, data_id: &DataIdentifier) -> Result<Data, Error>

Retrieve data we have on disk, that is also marked valid in the data chain.

fn put_data(&mut self, data: &Data) -> Result<BlockIdentifier, Error>

Add received data, return Result false if we do not have the corresponding valid Vote for this data. Will return a BlockIDentifier from us that we must use to create a Vote to send to peers. We also must add this Vote ourselves to this container. This is done in this manner to prevent coupling with keypairs etc.

Versioned ledger structured data will be Put and paid for

fn post_data(&mut self, data: &Data) -> Result<BlockIdentifier, Error>

Handle POST data This is a call that will only handle structured data

Will not accept versioned ledger based structuredData !

fn delete_data(&mut self,
data_id: &DataIdentifier,
_sigs: &[Signature])
-> Result<BlockIdentifier, Error>

Handle Delete data Unless ledger bit is set

fn provable_chain(&self, group_size: usize) -> DataChain

Return a chain for which we hold all of the data. Restricted to data that has a corresponding valid Block.

fn purge_disk(&mut self) -> Result<(), Error>

Remove any data on disk that we do not have a valid Block for

fn merge_chain(&mut self, chain: &mut DataChain)

Confirm and merge a DataChain transmitted to us. This will trim (purge invalid) exsiting entries then merge valid entries. May be used to create a new chain from given chains on node startup.

fn trust_level(&self, node: &PublicKey) -> usize

How many network events a given proover has been involved in (proover == node) First missed event stops the count

fn required_data(&self) -> Vec<BlockIdentifier>

Find any data we should have but are missing, given our current chain. The output of this gives an identifier we should send to other nodes to get the data. This is not a DataIdentifier as expected as this contains the hash we know the data must match.

fn max_space(&self) -> u64

Max space avilable for disk storage (as set by user)

fn used_space(&self) -> u64

Disk used so far.