Crate data_chain

Data Chain

This crate holds three modules that together allow nodes in a decentralised network to manage data and history of data and nodes in a container that in itself is decentralised.

The modules here include the ability to cryptographically lock history of data and nodes in a chain. There also exists the mechnism to persist these chains over sessions along with a copy of the data itself (the ChunkStore). Data representation is via the Data module.

Example

Basic usage

TBD

Github repository

Reexports

pub use chain::{DataChain, NodeBlock, Proof, create_link_descriptor};

pub use data::{ImmutableData, MAX_BYTES, PlainData, StructuredData};

Modules

chain
A block is a type that contains a BlockIdentifier and a Proof. These can be data blocks or links. When enough blocks (NodeBlocks) are received from other nodes a block becomes valid. This is a cetnral type to the security of republishable data on the network.

data
Data types Immutable data (name == hash of content) and Structured data (constant name with editable contents and ownership changes) this type is signed by at least one owner and can be a ledger type (never deleted)

error
Error types for this crate

secured_data
API This is the entry point to this crate and allows the crate to be used as a secured data store for all data types mentioned above.

sha3
sha3 (keccak)

Structs

Block
Used to validate chain Block can be a data item or a chain link.

Enums

BlockIdentifier
Data identifiers for use in a data Chain. The hash of each data type is available to ensure there is no confusion over the validity of any data presented by this chain

Data
Data types handled in a SAFE

DataIdentifier
An identifier to address a data chunk.