Crate data_chain

data_chain

Data blocks can be chained to provide verifiable assuredness that they contain network valid data and not injected.

A chain nay look like

`link` - all current group members - cryptographically secured as valid

 data - DataIdentifiers interspersed in a chain of links

 data - each block signed by a majority of previous link

 ....

`link` - As churn events occur new links are created

`link` - each link will have a majority
         (usually n - 1 actually) members of previous link

The chain, when presented to the current close_group (the group identified in last link) can be validated as holding data that has been agreed by the network over time to exist, A chain provides cryptographic proofs of data and group memberships over time.

The link is a group agreement chain component which is created by sorting the closest nodes to a network node (Address) and sending this hash, signed to that node. The recipient will then receive these, NodeBlocks and create the chain link. This link will allow the agreed members of the group so sign DataBlocks for the chain If a majority of the link members sign the data, it is validly in the chain. On group membership changes, a new link is constructed in the chain and the process repeats. A chain can split and nodes will maintain several chains, dependent on data overlaps with the chains links. The link identifier is the hash of all group members that contains the current node.

Containers of Chain Links only may also be maintained in groups to prove historic memberships of a network. It is not well enough understoof the validity of this action, but it may prove valuable in the event of network restarts.

Modules

block
A block is a type that contains a BlockIdentifier and a Proof. These can be data blocks or links

data_chain
A container of links (validated group membership blocks) and normal blocks (data elements)

error
Error types for this crate

node_block
A node block is a partial block, sent by group members to each other to create a Block

Constants

GROUP_SIZE
Required for consensus agreements. We should work this out though, Magic numbers :-(