Module data_chain::chain

A block is a type that contains a BlockIdentifier and a Proof. These can be data blocks or links

Reexports

pub use chain::node_block::{NodeBlock, create_link_descriptor};

pub use chain::data_chain::DataChain;

Modules

data_chain
A container of links (validated group membership blocks) and normal blocks (data elements)

node_block
A node block is a partial block, sent by group members to each other to create a Block

Enums

BlockIdentifier
Data identifiers for use in a data Chain. The hash of each data type is available to ensure there is no confusion over the validity of any data presented by this chain