Struct data_chain::secured_data::SecuredData

pub struct SecuredData {
    // some fields omitted
}

API for data based operations.

Methods

impl SecuredData

fn create_in_path(path: PathBuf, max_disk_space: u64) -> Result<SecuredData, Error>

Construct new data container

fn from_path(path: PathBuf, max_disk_space: u64) -> Result<SecuredData, Error>

Open an existing container from path

fn clear_disk(&self, path: &Path) -> Result<(), Error>

remove all disk based data

fn chain(&self) -> Arc<Mutex<DataChain>>

Access to DataChain

fn add_node_block(&mut self, nb: NodeBlock) -> Option<(BlockIdentifier, bool)>

Add a NodeBlock from another node If block is not a link and is valid wil return BlockIdentifier plus a bool to represent whether we have the data when the block is valid

fn get(&self, data_id: &DataIdentifier) -> Result<Data, Error>

Retrieve data we have on disk, that is also marked valid in the data chain.

fn put_data(&mut self, data: &Data) -> Result<BlockIdentifier, Error>

Add received data, return Result false if we do not have the corresponding valid NodeBlock for this data. Will return a BlockIDentifier from us that we must use to create a NodeBlock to send to peers. We also must add this NodeBlock ourselves to this container. This is done in this manner to prevent coupling with keypairs etc.

Versioned ledger structured data will be Put and paid for

fn post_data(&mut self, data: &Data) -> Result<BlockIdentifier, Error>

Handle POST data This is a call that will only handle structured data

Will not accept versioned ledger based structuredData !

fn delete_data(&mut self, data_id: &DataIdentifier) -> Result<BlockIdentifier, Error>

Handle Delete data :- will ignore the ledger bit

fn provable_chain(&self) -> DataChain

Return a chain for which we hold all of the data. Restricted to data that has a corresponding valid Block.

fn purge_disk(&mut self) -> Result<(), Error>

Remove any data on disk that we do not have a valid Block for

fn merge_chain(&mut self, chain: &mut DataChain)

Confirm and merge a DataChain transmitted to us. This will trim (purge invalid) exsiting entries then merge valid entries. May be used to create a new chain from given chains on node startup.

fn trust_level(&self, node: &PublicKey) -> usize

How many network events a given proover has been involved in (proover == node) First missed event stops the count

## Suggestion

We subtract from this score, blocks where the node has not responded (yet) If trust goes negative then the node may be killed from the network

fn required_data(&self) -> Vec<BlockIdentifier>

Find any data we should have, given our current chain

fn trim_all_blocks_and_data(&mut self)

Trim chain to previous common leading bits (previous vertice in binary tree) If our leading bits in group are 10111 then it will trim any 10110 data & links. Note this does not trim the chain to Genesis, that is maintained.